cleanup, reorganization, minor debugging
